--- 2020_Humanoid/ros_package/rbiz_autorace_monitor/CMakeLists.txt
@@ -1,58 +1,0 @@
-##############################################################################
-# CMake
-##############################################################################
-
-cmake_minimum_required(VERSION 2.8.3)
-project(rbiz_autorace_monitor)
-add_compile_options(-std=c++11)
-##############################################################################
-# Catkin
-##############################################################################
-
-# qt_build provides the qt cmake glue, roscpp the comms for a default talker
-find_package(catkin REQUIRED COMPONENTS 
-  qt_build
-  roscpp
-  rbiz_autorace_msgs
-)
-include_directories(${catkin_INCLUDE_DIRS})
-# Use this to define what the package will export (e.g. libs, headers).
-# Since the default here is to produce only a binary, we don't worry about
-# exporting anything.
-catkin_package(
-  CATKIN_DEPENDS rbiz_autorace_msgs
-)
-
-##############################################################################
-# Qt Environment
-##############################################################################
-
-# this comes from qt_build's qt-ros.cmake which is automatically
-# included via the dependency call in package.xml
-rosbuild_prepare_qt4(QtCore QtGui) # Add the appropriate components to the component list here
-
-##############################################################################
-# Sections
-##############################################################################
-
-file(GLOB QT_FORMS RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} ui/*.ui)
-file(GLOB QT_RESOURCES RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} resources/*.qrc)
-file(GLOB_RECURSE QT_MOC RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} FOLLOW_SYMLINKS include/rbiz_autorace_monitor/*.hpp)
-
-QT4_ADD_RESOURCES(QT_RESOURCES_CPP ${QT_RESOURCES})
-QT4_WRAP_UI(QT_FORMS_HPP ${QT_FORMS})
-QT4_WRAP_CPP(QT_MOC_HPP ${QT_MOC})
-
-##############################################################################
-# Sources
-##############################################################################
-
-file(GLOB_RECURSE QT_SOURCES RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} FOLLOW_SYMLINKS src/*.cpp)
-
-##############################################################################
-# Binaries
-##############################################################################
-add_executable(rbiz_autorace_monitor ${QT_SOURCES} ${QT_RESOURCES_CPP} ${QT_FORMS_HPP} ${QT_MOC_HPP})
-add_dependencies(rbiz_autorace_monitor ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
-target_link_libraries(rbiz_autorace_monitor ${QT_LIBRARIES} ${catkin_LIBRARIES})
-install(TARGETS rbiz_autorace_monitor RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
