--- README.md
@@ -1,2 +1,0 @@
-# 2020 Autorace Referee [Taiwan]
-Modify from [ROBOTIS-GIT/autorace_referee](https://github.com/ROBOTIS-GIT/autorace_referee)
