+++ rbiz_autorace_monitor/package.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>rbiz_autorace_monitor</name>
+  <version>1.0.0</version>
+  <description>The autorace_monitor package</description>
+  <license>Apache 2.0</license>
+  <maintainer email="happy170215@hotmail.com">zhl017</maintainer>
+  <buildtool_depend>catkin</buildtool_depend>
+  <depend>roscpp</depend>
+  <depend>std_msgs</depend>
+  <depend>qtbase5-dev</depend>
+  <depend>qt5-qmake</depend>
+  <depend>libqt5-core</depend>
+  <depend>libqt5-gui</depend>
+  <depend>cmake_modules</depend>
+  <depend>rbiz_autorace_msgs</depend>
+</package>
