+++ rbiz_autorace_msgs/msg/SensorStateLevelCrossing.msg
