+++ rbiz_autorace_msgs/msg/SensorStateStopwatch.msg
