+++ rbiz_autorace_msgs/msg/SensorStateTrafficLight.msg
