+++ rbiz_autorace_msgs/msg/Stage.msg
